swupdate_preserve_files: also copy dead links
dead links don't pass [ -e ] test, so were considered as not existing on the
source incorrectly.
shell is horrible...

+++ scripts/post_rootfs.sh
@@ -151,8 +151,9 @@ overwrite_to_target() {
 	local dir
 
 	for file; do
-		# source file must exist
-		[ -e "$file" ] || continue
+		# source file must exist... being careful of symlinks
+		[ -L "$file" ] || [ -e "$file" ] || continue
+
 		dir="${file%/*}"
 		mkdir_p_target "$dir"
 		rm -rf "$TARGET/$f"
